update frame class to be more like potential class
